Reduce duplicated scripting for hats-demo build process
This CL reuses kokoro build container and trusted application build-for-test.sh to remove duplication. The docker build process removes dependency on root and will not modify chmod / chown of the files in the git repo. This is done through having uid/gid inside the container to match the host identity.

Bug: 389144936
Change-Id: I746ec34e197dd5400605b3c59c24396700fab201